+++ lib/moxl/src/Moxl/Stanza/Muclumbus.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Moxl\Stanza;
+
+class Muclumbus
+{
+    public static function search($keyword)
+    {
+        $dom = new \DOMDocument('1.0', 'UTF-8');
+        $search = $dom->createElementNS('https://xmlns.zombofant.net/muclumbus/search/1.0', 'search');
+
+        $set = $dom->createElementNS('http://jabber.org/protocol/rsm', 'set');
+        $set->appendChild($dom->createElement('max', 30));
+        $search->appendChild($set);
+
+        $x = $dom->createElementNS('jabber:x:data', 'x');
+        $x->setAttribute('type', 'submit');
+        $search->appendChild($x);
+
+        $field = $dom->createElement('field');
+        $field->setAttribute('var', 'FORM_TYPE');
+        $field->setAttribute('type', 'hidden');
+        $field->appendChild($dom->createElement('value', 'https://xmlns.zombofant.net/muclumbus/search/1.0#params'));
+        $x->appendChild($field);
+
+        $q = $dom->createElement('field');
+        $q->setAttribute('var', 'q');
+        $q->setAttribute('type', 'text-single');
+        $q->appendChild($dom->createElement('value', $keyword));
+        $x->appendChild($q);
+
+        $xml = \Moxl\API::iqWrapper($search, 'rodrigo.de.mucobedo@dreckshal.de', 'get');
+        \Moxl\API::request($xml);
+    }
+}